requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
uritemplate==4.1.1
wcwidth==0.2.9
